prov/efa: fix ineffective error check on ofi_get_page_size()
ofi_get_page_size() returns long and uses negative values to indicate
errors. Storing the result in size_t (unsigned) makes the <= 0 error
check dead code, since a negative value wraps to a large positive.

Change the variable type from size_t to long so the error check works
as intended.

Fixes: Coverity CID 501834

+++ prov/efa/src/efa_domain.c
@@ -299,7 +299,7 @@ int efa_domain_open(struct fid_fabric *fabric_fid, struct fi_info *info,
 			struct fid_mr *mr_fid;
 			uint64_t mr_flags = FI_READ | FI_WRITE;
 
-			size_t page_size = ofi_get_page_size();
+			long page_size = ofi_get_page_size();
 			if (page_size <= 0) {
 				EFA_WARN(FI_LOG_DOMAIN, "Failed to get page size\n");
 				goto err_free;
